Migrate webhooks settings
