+++ packages/webapp/src/pages/Environment/Settings/Show.tsx
@@ -5,9 +5,9 @@ import { ConnectUISettings } from './ConnectUISettings';
 import { DeprecatedSettings } from './Deprecated';
 import { Functions } from './Functions';
 import { General } from './General';
-import { Notifications } from './Notifications';
 import { SlackAlertsSettings } from './SlackAlerts';
 import { Telemetry } from './Telemetry';
+import { Notifications } from './Webhooks';
 import { useEnvironment } from '../../../hooks/useEnvironment';
 import { useTeam } from '../../../hooks/useTeam';
 import DashboardLayout from '../../../layout/DashboardLayout';
